[Refactor] 알림 presentation 책임 분리
