+++ apps/mobile/lib/main.dart
@@ -9,6 +9,7 @@ import 'package:flutter/services.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'accessible_design.dart';
+import 'app/app_components.dart';
 import 'app/app_bootstrap.dart';
 import 'app/app_dependencies.dart';
 import 'core/datapack/data_pack_metered_consent_gate.dart';
@@ -22,7 +23,6 @@ import 'features/home_widget/next_train_widget_repository.dart';
 import 'features/home_widget/next_train_widget_runtime.dart'
     as next_train_widget_runtime;
 import 'facility_report.dart';
-import 'facility_status.dart';
 import 'favorite_facility.dart';
 import 'features/realtime/realtime_repository.dart';
 import 'features/route_draft/application/route_draft_controller.dart';
@@ -30,6 +30,7 @@ import 'features/route_draft/domain/route_draft.dart';
 import 'features/mobility_profile/mobility_preset_labels.dart';
 import 'features/mobility_profile/mobility_preset_picker.dart';
 import 'features/mobility_profile/mobility_profile_policy.dart';
+import 'features/notifications/presentation/notification_inbox_screen.dart';
 import 'internal_route.dart';
 import 'legacy_credential_cleanup.dart';
 import 'network_map.dart';
@@ -44,6 +45,10 @@ import 'station_search.dart';
 import 'mobile_error_reporter.dart';
 import 'user_data_deletion.dart';
 
+export 'app/app_components.dart' show FeatureTile;
+export 'features/notifications/presentation/notification_inbox_screen.dart'
+    show NotificationInboxScreen;
+
 const defaultPushNotificationsEnabled = bool.fromEnvironment(
   'EASYSUBWAY_ENABLE_PUSH_NOTIFICATIONS',
   defaultValue: false,
@@ -52,11 +57,7 @@ const defaultDemoHomeDataEnabled = bool.fromEnvironment(
   'EASYSUBWAY_DEMO_HOME_DATA',
   defaultValue: false,
 );
-const _mainPagePadding = EdgeInsets.fromLTRB(20, 20, 20, 32);
-const _mainListPagePadding = EdgeInsets.fromLTRB(17, 18, 17, 32);
-const _appSectionTitlePadding = EdgeInsets.fromLTRB(1, 22, 1, 11);
 const _settingsPagePadding = EdgeInsets.fromLTRB(20, 16, 20, 32);
-const _mainThemeControlRadius = BorderRadius.all(Radius.circular(12));
 const _mainIconControlRadius = BorderRadius.all(Radius.circular(12));
 
 Future<void> main() async {
@@ -490,7 +491,7 @@ class EasySubwayApp extends StatelessWidget {
           style: FilledButton.styleFrom(
             minimumSize: const Size.fromHeight(EasySubwayTouchTarget.primary),
             shape: const RoundedRectangleBorder(
-              borderRadius: _mainThemeControlRadius,
+              borderRadius: mainThemeControlRadius,
             ),
             textStyle: const TextStyle(
               fontSize: 18,
@@ -509,7 +510,7 @@ class EasySubwayApp extends StatelessWidget {
               width: 1.5,
             ),
             shape: const RoundedRectangleBorder(
-              borderRadius: _mainThemeControlRadius,
+              borderRadius: mainThemeControlRadius,
             ),
             textStyle: const TextStyle(
               fontSize: 16,
@@ -2074,7 +2075,7 @@ class _HomeScreenState extends State<HomeScreen> with WidgetsBindingObserver {
     if (facilitiesFuture != null) {
       try {
         final facilities = await facilitiesFuture;
-        if (facilities.any(_isFacilityAlert)) {
+        if (facilities.any(isFacilityAlert)) {
           return true;
         }
       } catch (error, stackTrace) {
@@ -2196,502 +2197,6 @@ class _HomeNotificationButton extends StatelessWidget {
   }
 }
 
-class NotificationInboxScreen extends StatefulWidget {
-  const NotificationInboxScreen({
-    required this.favoriteFacilityRepository,
-    required this.reportRepository,
-    required this.notificationRepository,
-    required this.notificationPermissionProvider,
-    super.key,
-  });
-
-  final FavoriteFacilityRepository? favoriteFacilityRepository;
-  final FacilityReportRepository reportRepository;
-  final NotificationSettingsRepository? notificationRepository;
-  final NotificationPermissionProvider? notificationPermissionProvider;
-
-  @override
-  State<NotificationInboxScreen> createState() =>
-      _NotificationInboxScreenState();
-}
-
-class _NotificationInboxScreenState extends State<NotificationInboxScreen> {
-  late Future<List<_NotificationInboxItem>> _itemsFuture;
-
-  @override
-  void initState() {
-    super.initState();
-    _itemsFuture = _loadItemsForDisplay();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('알림'),
-        actions: [
-          if (widget.notificationRepository != null)
-            IconButton(
-              tooltip: '알림 설정',
-              onPressed: _openSettings,
-              icon: const Icon(Icons.settings_outlined),
-            ),
-        ],
-      ),
-      body: SafeArea(
-        child: FutureBuilder<List<_NotificationInboxItem>>(
-          future: _itemsFuture,
-          builder: (context, snapshot) {
-            final items = snapshot.data ?? const <_NotificationInboxItem>[];
-            return RefreshIndicator(
-              onRefresh: () async {
-                final next = _loadItemsForDisplay();
-                setState(() {
-                  _itemsFuture = next;
-                });
-                try {
-                  await next;
-                } catch (error, stackTrace) {
-                  reportMobileError(
-                    error,
-                    stackTrace,
-                    context: '알림함 새로고침 중 예외가 발생했습니다.',
-                  );
-                }
-              },
-              child: ListView(
-                physics: const AlwaysScrollableScrollPhysics(),
-                padding: _mainListPagePadding,
-                children: [
-                  if (snapshot.connectionState != ConnectionState.done)
-                    const LinearProgressIndicator(minHeight: 3),
-                  if (snapshot.hasError)
-                    _HomeStateCard(
-                      key: const Key('notificationInboxErrorState'),
-                      icon: Icons.error_outline,
-                      title: '알림을 불러오지 못했어요',
-                      subtitle: '잠시 후 다시 시도해 주세요.',
-                      actionLabel: '다시 시도',
-                      onAction: () {
-                        setState(() {
-                          _itemsFuture = _loadItemsForDisplay();
-                        });
-                      },
-                    )
-                  else if (items.isEmpty)
-                    Padding(
-                      padding: const EdgeInsets.only(top: 80),
-                      child: Column(
-                        children: [
-                          const Icon(
-                            Icons.notifications_none,
-                            size: 44,
-                            color: EasySubwayAccessibleColors.mutedText,
-                          ),
-                          const SizedBox(height: 12),
-                          Text(
-                            '새 알림이 없습니다',
-                            textAlign: TextAlign.center,
-                            style: const TextStyle(
-                              color: EasySubwayAccessibleColors.text,
-                              fontSize: 16,
-                              fontWeight: FontWeight.w700,
-                            ),
-                          ),
-                          const SizedBox(height: 6),
-                          Text(
-                            '즐겨찾기 시설과 제보 상태가 바뀌면 여기에서 볼 수 있어요.',
-                            textAlign: TextAlign.center,
-                            style: const TextStyle(
-                              color: EasySubwayAccessibleColors.mutedText,
-                              fontSize: 13,
-                              fontWeight: FontWeight.w500,
-                              height: 1.4,
-                            ),
-                          ),
-                        ],
-                      ),
-                    )
-                  else ...[
-                    _NotificationInboxChips(items: items),
-                    for (final item in items) _NotificationInboxRow(item: item),
-                  ],
-                ],
-              ),
-            );
-          },
-        ),
-      ),
-    );
-  }
-
-  Future<List<_NotificationInboxItem>> _loadItemsForDisplay() {
-    final next = _loadItems();
-    unawaited(
-      next.catchError((Object error, StackTrace stackTrace) {
-        return const <_NotificationInboxItem>[];
-      }),
-    );
-    return next;
-  }
-
-  Future<List<_NotificationInboxItem>> _loadItems() async {
-    final items = <_NotificationInboxItem>[];
-    var loadFailed = false;
-    final favoriteFacilityRepository = widget.favoriteFacilityRepository;
-    if (favoriteFacilityRepository != null) {
-      try {
-        final facilities = await favoriteFacilityRepository
-            .listFavoriteFacilities();
-        for (final facility in facilities.where(_isFacilityAlert)) {
-          items.add(_NotificationInboxItem.facility(facility));
-        }
-      } catch (error, stackTrace) {
-        loadFailed = true;
-        reportMobileError(
-          error,
-          stackTrace,
-          context: '알림함 즐겨찾기 시설 상태를 불러오는 중 예외가 발생했습니다.',
-        );
-      }
-    }
-
-    try {
-      final reports = await widget.reportRepository.listMyReports();
-      for (final report in reports) {
-        items.add(_NotificationInboxItem.report(report));
-      }
-    } catch (error, stackTrace) {
-      loadFailed = true;
-      reportMobileError(
-        error,
-        stackTrace,
-        context: '알림함 제보 상태를 불러오는 중 예외가 발생했습니다.',
-      );
-    }
-    if (items.isEmpty && loadFailed) {
-      throw StateError('notification inbox load failed');
-    }
-    return items;
-  }
-
-  void _openSettings() {
-    final repository = widget.notificationRepository;
-    if (repository == null) {
-      return;
-    }
-    Navigator.of(context).push(
-      MaterialPageRoute<void>(
-        builder: (_) => NotificationSettingsScreen(
-          repository: repository,
-          notificationPermissionProvider: widget.notificationPermissionProvider,
-        ),
-      ),
-    );
-  }
-}
-
-class _NotificationInboxItem {
-  const _NotificationInboxItem({
-    required this.icon,
-    required this.title,
-    required this.subtitle,
-    required this.semanticLabel,
-    required this.kind,
-    this.report,
-    this.severity = FacilityStatusSeverity.normal,
-    this.actionLabel = '',
-  });
-
-  factory _NotificationInboxItem.facility(FavoriteFacility facility) {
-    final name = facility.name.trim().isEmpty
-        ? facility.typeLabel
-        : facility.name;
-    return _NotificationInboxItem(
-      icon: _facilityIcon(facility.type),
-      title: '${facility.stationLabel} $name',
-      subtitle:
-          '${facility.severityLabel} · ${facility.typeLabel} ${facility.statusLabel}',
-      semanticLabel:
-          '${facility.stationLabel} $name, ${facility.typeLabel} ${facility.statusLabel}, ${facility.severityLabel}, ${facility.updatedLabel}, ${facility.dataSourceLabel}, ${facility.nextActionLabel}',
-      kind: '시설',
-      severity: facility.statusPresentation.severity,
-      actionLabel: facility.nextActionLabel,
-    );
-  }
-
-  factory _NotificationInboxItem.report(FacilityReportResult report) {
-    return _NotificationInboxItem(
-      icon: Icons.report_outlined,
-      title: '제보 ${report.statusLabel}',
-      subtitle: '제보 번호 ${report.displayReceiptCode}',
-      semanticLabel:
-          '제보 ${report.statusLabel}, 제보 번호 ${report.displayReceiptCode}',
-      kind: '제보',
-      report: report,
-    );
-  }
-
-  final IconData icon;
-  final String title;
-  final String subtitle;
-  final String semanticLabel;
-  final String kind;
-  final FacilityReportResult? report;
-  final FacilityStatusSeverity severity;
-  final String actionLabel;
-}
-
-class _NotificationInboxChips extends StatelessWidget {
-  const _NotificationInboxChips({required this.items});
-
-  final List<_NotificationInboxItem> items;
-
-  @override
-  Widget build(BuildContext context) {
-    final facilityCount = items.where((item) => item.kind == '시설').length;
-    final reportCount = items.length - facilityCount;
-    final parts = <String>[
-      '전체 ${items.length}',
-      if (facilityCount > 0) '시설 $facilityCount',
-      if (reportCount > 0) '제보 $reportCount',
-    ];
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 4, top: 4),
-      child: Text(
-        parts.join('  ·  '),
-        style: const TextStyle(
-          color: EasySubwayAccessibleColors.mutedText,
-          fontSize: 13,
-          fontWeight: FontWeight.w600,
-        ),
-      ),
-    );
-  }
-}
-
-class _NotificationInboxRow extends StatelessWidget {
-  const _NotificationInboxRow({required this.item});
-
-  final _NotificationInboxItem item;
-
-  @override
-  Widget build(BuildContext context) {
-    void open() {
-      final report = item.report;
-      if (report == null) {
-        return;
-      }
-      Navigator.of(context).push(
-        MaterialPageRoute<void>(
-          builder: (_) => MyFacilityReportDetailScreen(report: report),
-        ),
-      );
-    }
-
-    final accent = _facilitySeverityAccent(item.severity);
-    final row = Container(
-      padding: const EdgeInsets.symmetric(vertical: 14),
-      decoration: const BoxDecoration(
-        border: Border(
-          bottom: BorderSide(color: EasySubwayAccessibleColors.line),
-        ),
-      ),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(top: 1),
-            child: Icon(item.icon, color: accent.iconColor, size: 22),
-          ),
-          const SizedBox(width: 12),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Expanded(
-                      child: Text(
-                        item.title,
-                        style: const TextStyle(
-                          color: EasySubwayAccessibleColors.text,
-                          fontSize: 15,
-                          fontWeight: FontWeight.w700,
-                          height: 1.25,
-                        ),
-                      ),
-                    ),
-                    const SizedBox(width: 8),
-                    Text(
-                      item.kind,
-                      style: const TextStyle(
-                        color: EasySubwayAccessibleColors.mutedText,
-                        fontSize: 12,
-                        fontWeight: FontWeight.w600,
-                      ),
-                    ),
-                  ],
-                ),
-                const SizedBox(height: 3),
-                Text(
-                  item.subtitle,
-                  style: TextStyle(
-                    color: accent.iconColor,
-                    fontSize: 12,
-                    fontWeight: FontWeight.w500,
-                    height: 1.4,
-                  ),
-                ),
-                if (item.actionLabel.isNotEmpty) ...[
-                  const SizedBox(height: 6),
-                  Text(
-                    item.actionLabel,
-                    style: const TextStyle(
-                      color: EasySubwayAccessibleColors.text,
-                      fontSize: 13,
-                      fontWeight: FontWeight.w600,
-                      height: 1.3,
-                    ),
-                  ),
-                ],
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-    if (item.report == null) {
-      return Semantics(
-        label: item.semanticLabel,
-        child: ExcludeSemantics(child: row),
-      );
-    }
-    return Semantics(
-      button: true,
-      label: item.semanticLabel,
-      onTap: open,
-      child: ExcludeSemantics(
-        child: InkWell(onTap: open, child: row),
-      ),
-    );
-  }
-}
-
-class _AppSectionTitle extends StatelessWidget {
-  const _AppSectionTitle({required this.title});
-
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: _appSectionTitlePadding,
-      child: Text(
-        title,
-        style: Theme.of(context).textTheme.titleMedium?.copyWith(
-          color: EasySubwayAccessibleColors.text,
-          fontWeight: FontWeight.w700,
-          height: 1.2,
-        ),
-      ),
-    );
-  }
-}
-
-class _FacilitySeverityAccent {
-  const _FacilitySeverityAccent({
-    required this.backgroundColor,
-    required this.borderColor,
-    required this.iconColor,
-  });
-
-  final Color backgroundColor;
-  final Color borderColor;
-  final Color iconColor;
-}
-
-_FacilitySeverityAccent _facilitySeverityAccent(
-  FacilityStatusSeverity severity,
-) {
-  return switch (severity) {
-    FacilityStatusSeverity.blocked => const _FacilitySeverityAccent(
-      backgroundColor: EasySubwayAccessibleColors.redSoft,
-      borderColor: EasySubwayAccessibleColors.red,
-      iconColor: EasySubwayAccessibleColors.red,
-    ),
-    FacilityStatusSeverity.caution => const _FacilitySeverityAccent(
-      backgroundColor: EasySubwayAccessibleColors.amberSoft,
-      borderColor: EasySubwayAccessibleColors.amberBorder,
-      iconColor: EasySubwayAccessibleColors.amber,
-    ),
-    FacilityStatusSeverity.needsInfo => const _FacilitySeverityAccent(
-      backgroundColor: Colors.white,
-      borderColor: EasySubwayAccessibleColors.needsInfo,
-      iconColor: EasySubwayAccessibleColors.needsInfo,
-    ),
-    FacilityStatusSeverity.normal => const _FacilitySeverityAccent(
-      backgroundColor: Colors.white,
-      borderColor: EasySubwayAccessibleColors.line,
-      iconColor: EasySubwayAccessibleColors.mintDark,
-    ),
-  };
-}
-
-bool _isFacilityAlert(FavoriteFacility facility) {
-  return facility.needsAttention;
-}
-
-IconData _facilityIcon(String type) {
-  return switch (type) {
-    'ELEVATOR' => Icons.elevator_outlined,
-    'ESCALATOR' => Icons.escalator_warning_outlined,
-    'WHEELCHAIR_LIFT' => Icons.accessible_forward,
-    'RAMP' => Icons.accessible,
-    'ACCESSIBLE_TOILET' || 'TOILET' => Icons.wc_outlined,
-    _ => Icons.warning_amber_outlined,
-  };
-}
-
-class _HomeStateCard extends StatelessWidget {
-  const _HomeStateCard({
-    required this.icon,
-    required this.title,
-    required this.subtitle,
-    this.actionLabel,
-    this.onAction,
-    super.key,
-  });
-
-  final IconData icon;
-  final String title;
-  final String subtitle;
-  final String? actionLabel;
-  final VoidCallback? onAction;
-
-  @override
-  Widget build(BuildContext context) {
-    final actionLabel = this.actionLabel;
-    return _AppCard(
-      showBorder: true,
-      child: AccessibleStateCard(
-        icon: icon,
-        title: title,
-        subtitle: subtitle,
-        actions: [
-          if (actionLabel != null && onAction != null)
-            OutlinedButton.icon(
-              onPressed: onAction,
-              icon: const Icon(Icons.refresh),
-              label: Text(actionLabel),
-            ),
-        ],
-      ),
-    );
-  }
-}
-
 class _HomeSavedRouteCard extends StatelessWidget {
   const _HomeSavedRouteCard({
     required this.route,
@@ -2756,7 +2261,7 @@ class _HomeSavedRouteCard extends StatelessWidget {
         ),
       ),
     );
-    return _AppCard(
+    return AppCard(
       showBorder: true,
       child: Row(
         children: [
@@ -2809,94 +2314,6 @@ class _HomeMiniBadge extends StatelessWidget {
         height: 1.2,
       ),
       padding: const EdgeInsets.symmetric(horizontal: 2),
-    );
-  }
-}
-
-class _AppCard extends StatelessWidget {
-  const _AppCard({
-    required this.child,
-    this.backgroundColor = Colors.white,
-    this.borderColor = EasySubwayAccessibleColors.line,
-    this.showBorder = true,
-  });
-
-  final Widget child;
-  final Color backgroundColor;
-  final Color borderColor;
-  final bool showBorder;
-
-  @override
-  Widget build(BuildContext context) {
-    // 최소 그림자 원칙: 그림자 대신 얇은 보더로 카드를 구분한다.
-    return Card(
-      margin: EdgeInsets.zero,
-      color: backgroundColor,
-      elevation: 0,
-      shadowColor: EasySubwayAccessibleColors.cardShadow,
-      shape: RoundedRectangleBorder(
-        side: showBorder ? BorderSide(color: borderColor) : BorderSide.none,
-        borderRadius: BorderRadius.circular(20),
-      ),
-      clipBehavior: Clip.antiAlias,
-      child: Padding(padding: const EdgeInsets.all(16), child: child),
-    );
-  }
-}
-
-class _AppInfoRow extends StatelessWidget {
-  const _AppInfoRow({
-    required this.icon,
-    required this.iconColor,
-    required this.title,
-    this.subtitle,
-  });
-
-  final IconData icon;
-  final Color iconColor;
-  final String title;
-  final String? subtitle;
-
-  @override
-  Widget build(BuildContext context) {
-    final subtitle = this.subtitle;
-    final leading = SizedBox(
-      width: 32,
-      height: 32,
-      child: Center(child: Icon(icon, color: iconColor, size: 22)),
-    );
-    final content = Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          title,
-          style: const TextStyle(
-            color: EasySubwayAccessibleColors.text,
-            fontSize: 15,
-            fontWeight: FontWeight.w700,
-            height: 1.25,
-          ),
-        ),
-        if (subtitle != null && subtitle.isNotEmpty) ...[
-          const SizedBox(height: 3),
-          Text(
-            subtitle,
-            style: const TextStyle(
-              color: EasySubwayAccessibleColors.mutedText,
-              fontSize: 12,
-              fontWeight: FontWeight.w500,
-              height: 1.4,
-            ),
-          ),
-        ],
-      ],
-    );
-    return Row(
-      children: [
-        leading,
-        const SizedBox(width: 12),
-        Expanded(child: content),
-      ],
     );
   }
 }
@@ -3345,12 +2762,12 @@ class _FavoriteHomeScreenState extends State<FavoriteHomeScreen> {
               },
               child: ListView(
                 physics: const AlwaysScrollableScrollPhysics(),
-                padding: _mainListPagePadding,
+                padding: mainListPagePadding,
                 children: [
                   if (snapshot.connectionState != ConnectionState.done)
                     const LinearProgressIndicator(minHeight: 3),
                   if (hasError)
-                    _HomeStateCard(
+                    HomeStateCard(
                       key: const Key('favoriteHomeErrorState'),
                       icon: Icons.error_outline,
                       title: '즐겨찾기를 불러오지 못했어요',
@@ -3365,8 +2782,8 @@ class _FavoriteHomeScreenState extends State<FavoriteHomeScreen> {
                   else if (data.isEmpty)
                     const Padding(
                       padding: EdgeInsets.only(top: 16),
-                      child: _AppCard(
-                        child: _AppInfoRow(
+                      child: AppCard(
+                        child: AppInfoRow(
                           icon: Icons.bookmark_border,
                           iconColor: EasySubwayAccessibleColors.mutedText,
                           title: '즐겨찾기한 항목이 없습니다',
@@ -3378,7 +2795,7 @@ class _FavoriteHomeScreenState extends State<FavoriteHomeScreen> {
                     // 카테고리 카드·개수 없이 저장한 항목을 바로 나열한다. 섹션
                     // 헤더는 해당 항목이 있을 때만 보여준다(#1569).
                     if (data.stations.isNotEmpty) ...[
-                      const _AppSectionTitle(title: '역'),
+                      const AppSectionTitle(title: '역'),
                       for (final station in data.stations)
                         _FavoriteHomeStationRow(
                           station: station,
@@ -3388,7 +2805,7 @@ class _FavoriteHomeScreenState extends State<FavoriteHomeScreen> {
                         ),
                     ],
                     if (data.routes.isNotEmpty) ...[
-                      const _AppSectionTitle(title: '경로'),
+                      const AppSectionTitle(title: '경로'),
                       for (final route in data.routes)
                         Padding(
                           padding: const EdgeInsets.only(bottom: 8),
@@ -3402,7 +2819,7 @@ class _FavoriteHomeScreenState extends State<FavoriteHomeScreen> {
                         ),
                     ],
                     if (data.facilities.isNotEmpty) ...[
-                      const _AppSectionTitle(title: '시설'),
+                      const AppSectionTitle(title: '시설'),
                       for (final facility in data.facilities)
                         _FavoriteHomeFacilityRow(
                           facility: facility,
@@ -3722,7 +3139,7 @@ class SupportAccessScreen extends StatelessWidget {
       appBar: AppBar(title: const Text('도움말·문의')),
       body: SafeArea(
         child: ListView(
-          padding: _mainPagePadding,
+          padding: mainPagePadding,
           children: [
             const _SupportSectionTitle(title: '내 정보와 개인정보'),
             _SupportGroupCard(
@@ -4122,7 +3539,7 @@ class _UserDataDeletionScreenState extends State<UserDataDeletionScreen> {
       ),
       body: SafeArea(
         child: ListView(
-          padding: _mainPagePadding,
+          padding: mainPagePadding,
           children: [
             Semantics(
               header: true,
@@ -4259,9 +3676,9 @@ class UserDataDeletionResultScreen extends StatelessWidget {
       ),
       body: SafeArea(
         child: ListView(
-          padding: _mainPagePadding,
+          padding: mainPagePadding,
           children: [
-            _AppCard(
+            AppCard(
               backgroundColor: EasySubwayAccessibleColors.surface,
               borderColor: EasySubwayAccessibleColors.line,
               child: Column(
@@ -4289,9 +3706,9 @@ class UserDataDeletionResultScreen extends StatelessWidget {
               ),
             ),
             const SizedBox(height: 12),
-            const _AppCard(
+            const AppCard(
               showBorder: true,
-              child: _AppInfoRow(
+              child: AppInfoRow(
                 icon: Icons.map_outlined,
                 iconColor: EasySubwayAccessibleColors.primary,
                 title: '노선도와 역 정보는 계속 이용할 수 있어요',
@@ -4556,10 +3973,10 @@ class _DataSourceAttributionScreenState
               builder: (context, snapshot) {
                 if (snapshot.hasError) {
                   return ListView(
-                    padding: _mainPagePadding,
+                    padding: mainPagePadding,
                     children: const [
-                      _AppCard(
-                        child: _AppInfoRow(
+                      AppCard(
+                        child: AppInfoRow(
                           icon: Icons.error_outline,
                           iconColor: EasySubwayAccessibleColors.amber,
                           title: '자료 제공 정보를 불러오지 못했어요',
@@ -4580,10 +3997,10 @@ class _DataSourceAttributionScreenState
                     .cast<Map<String, Object?>>()
                     .toList(growable: false);
                 return ListView(
-                  padding: _mainPagePadding,
+                  padding: mainPagePadding,
                   children: [
-                    const _AppCard(
-                      child: _AppInfoRow(
+                    const AppCard(
+                      child: AppInfoRow(
                         icon: Icons.fact_check_outlined,
                         iconColor: EasySubwayAccessibleColors.amber,
                         title: '현재 앱 표시',
@@ -4593,11 +4010,11 @@ class _DataSourceAttributionScreenState
                             '지도와 길·시설 안내는 공식·공개 자료를 바탕으로 해요. 상록수·사당역은 현장 확인까지 마쳤고, 나머지 역은 자료를 바탕으로 안내해요.',
                       ),
                     ),
-                    const _AppSectionTitle(title: '지도 표시용 asset'),
+                    const AppSectionTitle(title: '지도 표시용 asset'),
                     for (final map in maps) _AttributionCard.map(map, manifest),
-                    const _AppSectionTitle(title: '데이터 품질 Level'),
-                    const _AppCard(
-                      child: _AppInfoRow(
+                    const AppSectionTitle(title: '데이터 품질 Level'),
+                    const AppCard(
+                      child: AppInfoRow(
                         icon: Icons.verified_outlined,
                         iconColor: EasySubwayAccessibleColors.mintDark,
                         title: 'Level 1-4 품질 기준',
@@ -4605,8 +4022,8 @@ class _DataSourceAttributionScreenState
                             'Level 1은 역·노선 수, Level 2는 필수 시설 근거, Level 3은 운행상태와 최신 여부, Level 4는 현장 또는 운영기관이 확인한 쉬운 길을 봐요.',
                       ),
                     ),
-                    const _AppCard(
-                      child: _AppInfoRow(
+                    const AppCard(
+                      child: AppInfoRow(
                         icon: Icons.analytics_outlined,
                         iconColor: EasySubwayAccessibleColors.amber,
                         title: '품질 지표',
@@ -4614,7 +4031,7 @@ class _DataSourceAttributionScreenState
                             '필수 시설 근거 비율, 운행상태 확인 비율, 최신 정보 비율, 확인된 쉬운 길 비율, 현장 확인 경로 비율을 함께 확인해요.',
                       ),
                     ),
-                    const _AppSectionTitle(title: '경로·시설 안내용 데이터'),
+                    const AppSectionTitle(title: '경로·시설 안내용 데이터'),
                     for (final source in sources)
                       _AttributionCard.source(source),
                   ],
@@ -4703,7 +4120,7 @@ class _AttributionCard extends StatelessWidget {
       padding: const EdgeInsets.only(bottom: 12),
       child: Semantics(
         label: semanticLabel,
-        child: _AppCard(
+        child: AppCard(
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
@@ -4944,59 +4361,4 @@ Uri? _mailtoUri(String value, String subject) {
     path: email,
     queryParameters: {'subject': subject},
   );
-}
-
-class FeatureTile extends StatelessWidget {
-  const FeatureTile({
-    required this.icon,
-    required this.title,
-    required this.semanticLabel,
-    super.key,
-  });
-
-  final IconData icon;
-  final String title;
-  final String semanticLabel;
-
-  @override
-  Widget build(BuildContext context) {
-    final colorScheme = Theme.of(context).colorScheme;
-
-    return MergeSemantics(
-      child: Semantics(
-        label: semanticLabel,
-        child: ExcludeSemantics(
-          child: Card(
-            margin: const EdgeInsets.only(bottom: 12),
-            color: Colors.white,
-            elevation: 0,
-            shape: RoundedRectangleBorder(
-              borderRadius: _mainThemeControlRadius,
-              side: const BorderSide(color: EasySubwayAccessibleColors.line),
-            ),
-            child: Padding(
-              padding: const EdgeInsets.all(16),
-              child: Row(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  Icon(icon, color: colorScheme.primary, size: 32),
-                  const SizedBox(width: 16),
-                  Expanded(
-                    child: Text(
-                      title,
-                      style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                        color: EasySubwayAccessibleColors.text,
-                        fontWeight: FontWeight.w700,
-                        height: 1.35,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
 }
